pyproject.toml
[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[project]
name = "zotero-tracker"
version = "0.1.0"
description = "Daily arXiv digest matched to your Zotero library (Markdown email)."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "dotenv>=0.9.9",
    "feedparser>=6.0.11",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "numpy>=1.26.0",
    "openai>=2.24.0",
    "pyzotero>=1.10.0",
    "scikit-learn>=1.7.1",
    "tiktoken>=0.8.0",
    "tqdm>=4.66.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:hydra"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.uv]
package = true

[project.scripts]
zotero-tracker = "zotero_tracker.main:main"
